finish view api

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async (req, res) => {
@@ -5,12 +6,15 @@ export const home = async (req, res) => {
   //   return res.render("home", { pageTitle: "Home", videos });
   // }); 콜백 이용한 Video.find
 
-  const videos = await Video.find({}).sort({ createdAt: "desc" });
+  const videos = await Video.find({})
+    .sort({ createdAt: "desc" })
+    .populate("owner");
   return res.render("home", { pageTitle: "Home", videos });
 };
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
+  // const owner = await User.findById(video.owner); 바보같은 방법!
 
   if (!video) {
     // 누군가가 존재하지 않은 video 페이지를 방문했을 때를 대비한 것.
@@ -20,16 +24,22 @@ export const watch = async (req, res) => {
 };
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const { _id } = req.session.user;
+
   const video = await Video.findById(id); // video object가 반드시 필요함.
 
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/"); //403: forbidden
   }
   return res.render("edit", { pageTitle: `Edit ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
+  const { _id } = req.session.user;
 
   const video = await Video.exists({
     _id: id,
@@ -37,6 +47,10 @@ export const postEdit = async (req, res) => {
 
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/"); //403: forbidden
   }
 
   await Video.findByIdAndUpdate(id, {
@@ -55,7 +69,9 @@ export const search = async (req, res) => {
     const regex = new RegExp(keyword, "gi");
     videos = await Video.find({
       title: { $regex: regex },
-    }).sort({ createdAt: "desc" });
+    })
+      .sort({ createdAt: "desc" })
+      .populate("owner");
   }
 
   res.render("search", { pageTitle: "Search", videos });
@@ -63,6 +79,17 @@ export const search = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const { _id } = req.session.user;
+
+  const video = await Video.findById(id);
+
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/"); //403: forbidden
+  }
 
   await Video.findByIdAndDelete(id);
 
@@ -74,19 +101,27 @@ export const getUpload = (req, res) => {
 };
 export const postUpload = async (req, res) => {
   const { title, description, hashtags } = req.body;
+  const { file } = req;
+  const { _id } = req.session.user;
 
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       // 커서를 갖다대면 Video.create는 프로미즈를 리턴한다는 것을 알 수 있다.
       title,
       description,
+      fileUrl: file.path,
       // createdAt: Date.now(), 변하는 값이 아닌데 매번 아렇게 쳐야하는 건 곤욕. 따라서 schema에 가서 default 값으로 설정.
       hashtags: Video.formatHashtags(hashtags),
       // meta: {
       //   views: 0,
       //   rating: 0,
       // },
+      owner: _id,
     });
+
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
 
     return res.redirect(`/`);
   } catch (error) {
@@ -95,4 +130,18 @@ export const postUpload = async (req, res) => {
       errorMessage: error.message,
     });
   }
+};
+
+export const registerView = async (req, res) => {
+  const { id } = req.params;
+
+  const video = await Video.findById(id);
+
+  if (!video) {
+    return res.sendStatus(404); // 얜 주소를 바꾸지도, (그래서 누구나 접근 가능) 템플릿을 렌더링하지도 않는 컨트롤러임. 백엔드에 정보를 전송하고 처리하는 것만 함. 따라서 이 경우에 백엔드가 할 수 있는 대답은 이게 전부.
+  }
+
+  video.meta.views += 1;
+  await video.save();
+  return res.sendStatus(200);
 };
